Automatically set reduce_range
