+++ test/quantization/quantize_/workflows/int8/test_int8_tensor.py
@@ -19,23 +19,17 @@ from torchao.quantization import (
     quantize_,
 )
 from torchao.quantization.granularity import PerGroup, PerRow, PerTensor
-from torchao.quantization.quant_primitives import MappingType, choose_qparams_affine
+from torchao.quantization.quant_primitives import MappingType
 from torchao.quantization.quantize_.common import (
     _choose_quant_func_and_quantize_tensor,
 )
 from torchao.quantization.quantize_.workflows.int8.int8_tensor import (
-    _FULL_QUANT_MAX,
-    _FULL_QUANT_MIN,
-    _REDUCED_QUANT_MAX,
-    _REDUCED_QUANT_MIN,
-    Int8Tensor,
     QuantizeTensorToInt8Kwargs,
 )
 from torchao.quantization.utils import compute_error, get_block_size
 from torchao.testing.model_architectures import ToyTwoLinearModel
 from torchao.testing.utils import TorchAOIntegrationTestCase, skip_if_xpu
 from torchao.utils import (
-    _cpu_is_vnni_supported,
     get_available_devices,
     is_ROCM,
     torch_version_at_least,
@@ -519,94 +513,6 @@ class TestInt8StaticQuant(TorchAOIntegrationTestCase):
         # Compute SQNR and make sure it's above 40
         sqnr = compute_error(output_baseline, output)
         self.assertGreater(sqnr, 40, f"SQNR too low: {sqnr}")
-
-
-@common_utils.instantiate_parametrized_tests
-class TestInt8TensorCPU(TorchAOIntegrationTestCase):
-    @common_utils.parametrize("dtype", [torch.bfloat16, torch.float32])
-    @common_utils.parametrize("compile", [True, False])
-    @common_utils.parametrize("config_mode", ["dynamic", "static"])
-    @common_utils.parametrize(
-        "granularity",
-        [PerRow(), PerTensor(), (PerRow(), PerTensor()), (PerTensor(), PerRow())],
-    )
-    @common_utils.parametrize(
-        "act_mapping_type", [MappingType.SYMMETRIC, MappingType.ASYMMETRIC]
-    )
-    @common_utils.parametrize("reduce_range", [False, True])
-    def test_int8_tensor_cpu(
-        self, act_mapping_type, granularity, reduce_range, config_mode, compile, dtype
-    ):
-        device = "cpu"
-        if is_ROCM():
-            self.skipTest("Don't test CPU for ROCM version of torch")
-
-        torch.compiler.reset()
-
-        M, N, K = 64, 256, 256
-        input_tensor = torch.randn(M, K, dtype=dtype, device=device)
-        model = ToyTwoLinearModel(K, N, K, dtype=dtype, device=device).eval()
-        model_q = copy.deepcopy(model)
-
-        if config_mode == "dynamic":
-            config = Int8DynamicActivationInt8WeightConfig(
-                version=2,
-                granularity=granularity,
-                act_mapping_type=act_mapping_type,
-                reduce_range=reduce_range,
-            )
-        else:
-            act_granularity, _ = Int8Tensor._normalize_granularity(granularity)
-            quant_min, quant_max = (
-                (_REDUCED_QUANT_MIN, _REDUCED_QUANT_MAX)
-                if reduce_range
-                else (_FULL_QUANT_MIN, _FULL_QUANT_MAX)
-            )
-            block_size = get_block_size(input_tensor.shape, act_granularity)
-            act_quant_scale, act_quant_zero_point = choose_qparams_affine(
-                input=input_tensor,
-                mapping_type=act_mapping_type,
-                block_size=block_size,
-                target_dtype=torch.int8,
-                quant_min=quant_min,
-                quant_max=quant_max,
-                scale_dtype=dtype,
-                zero_point_dtype=torch.int8,
-                keepdim=True,
-                eps=torch.finfo(torch.float32).eps,
-            )
-            config = Int8StaticActivationInt8WeightConfig(
-                act_quant_scale=act_quant_scale,
-                act_quant_zero_point=act_quant_zero_point,
-                granularity=granularity,
-                act_mapping_type=act_mapping_type,
-                reduce_range=reduce_range,
-            )
-
-        quantize_(model_q, config)
-
-        _, weight_granularity = Int8Tensor._normalize_granularity(config.granularity)
-        if isinstance(weight_granularity, PerRow):
-            self.assertEqual(model_q.linear2.weight.scale.shape, (K, 1))
-        elif isinstance(weight_granularity, PerTensor):
-            self.assertEqual(model_q.linear2.weight.scale.shape, (1, 1))
-
-        self.assertEqual(model_q.linear2.weight.scale.ndim, 2)
-
-        if compile:
-            model_q = torch.compile(model_q, fullgraph=True)
-
-        output_fp = model(input_tensor)
-        output_quantized = model_q(input_tensor)
-
-        if reduce_range or _cpu_is_vnni_supported():
-            assert compute_error(output_fp, output_quantized) > 20, (
-                f"Quantization error is too high got a SQNR of {compute_error(output_fp, output_quantized)}"
-            )
-        else:
-            # At minimum, verify no NaN/Inf in output
-            assert not torch.isnan(output_quantized).any(), "Output contains NaN"
-            assert not torch.isinf(output_quantized).any(), "Output contains Inf"
 
 
 if __name__ == "__main__":
