+++ torchao/quantization/quant_api.py
@@ -932,9 +932,6 @@ class Int8DynamicActivationInt8WeightConfig(AOBaseConfig):
         set_inductor_config: bool = True - If True, adjusts `torchinductor` settings to recommended values
             for better performance with this quantization scheme.
         version (int): the version of the config
-        reduce_range (Optional[bool] = False): If True, both activation and weight int8 quantization use reduced range
-            [_REDUCED_QUANT_MIN, _REDUCED_QUANT_MAX] instead of full range
-            [_FULL_QUANT_MIN, _FULL_QUANT_MAX] to reduce overflow risk on platforms without VNNI instructions.
 
     Example:
 
@@ -949,7 +946,6 @@ class Int8DynamicActivationInt8WeightConfig(AOBaseConfig):
     ] = PerRow()
     set_inductor_config: bool = True
     version: int = 2
-    reduce_range: Optional[bool] = False
 
     def __post_init__(self):
         torch._C._log_api_usage_once(
@@ -974,9 +970,6 @@ class Int8DynamicActivationInt8WeightConfig(AOBaseConfig):
             "Please set it to MappingType.SYMMETRIC or "
             "MappingType.ASYMMETRIC."
         )
-        assert self.reduce_range in (True, False), (
-            "`reduce_range` must be True or False"
-        )
 
 
 def _int8_dynamic_activation_int8_weight_quantize_tensor(weight, config):
@@ -991,9 +984,7 @@ def _int8_dynamic_activation_int8_weight_quantize_tensor(weight, config):
         act_quant_kwargs=QuantizeTensorToInt8Kwargs(
             granularity=act_granularity,
             mapping_type=config.act_mapping_type,
-            reduce_range=config.reduce_range,
         ),
-        reduce_range=config.reduce_range,
     )
 
     return quantized_weight
@@ -1047,9 +1038,6 @@ class Int8StaticActivationInt8WeightConfig(AOBaseConfig):
         act_mapping_type (MappingType): The mapping type for activation quantization. SYMMETRIC and ASYMMETRIC are supported.
         set_inductor_config (bool): if True, adjusts `torchinductor` settings to recommended values.
         version (int): the version of the config
-        reduce_range (Optional[bool] = False): If True, both activation and weight int8 quantization use reduced range
-            [_REDUCED_QUANT_MIN, _REDUCED_QUANT_MAX] instead of full range
-            [_FULL_QUANT_MIN, _FULL_QUANT_MAX] to reduce overflow risk on platforms without VNNI instructions.
     """
 
     act_quant_scale: Optional[torch.Tensor] = None
@@ -1060,7 +1048,6 @@ class Int8StaticActivationInt8WeightConfig(AOBaseConfig):
     act_mapping_type: Optional[MappingType] = MappingType.SYMMETRIC
     set_inductor_config: bool = True
     version: int = 1
-    reduce_range: Optional[bool] = False
 
     def __post_init__(self):
         torch._C._log_api_usage_once(
@@ -1080,9 +1067,6 @@ class Int8StaticActivationInt8WeightConfig(AOBaseConfig):
             "Please set it to MappingType.SYMMETRIC or "
             "MappingType.ASYMMETRIC."
         )
-        assert self.reduce_range in (True, False), (
-            "`reduce_range` must be True or False"
-        )
 
     def get_act_quant_kwargs(self) -> QuantizeTensorToInt8Kwargs:
         """Get the activation quantization kwargs for static quantization.
@@ -1097,7 +1081,6 @@ class Int8StaticActivationInt8WeightConfig(AOBaseConfig):
         return QuantizeTensorToInt8Kwargs(
             granularity=act_granularity,
             mapping_type=self.act_mapping_type,
-            reduce_range=self.reduce_range,
         )
 
 
@@ -1128,11 +1111,9 @@ def _int8_static_activation_int8_weight_transform(
         act_quant_kwargs=QuantizeTensorToInt8Kwargs(
             granularity=activation_granularity,
             mapping_type=config.act_mapping_type,
-            reduce_range=config.reduce_range,
         ),
         act_quant_scale=config.act_quant_scale.detach(),
         act_quant_zero_point=act_quant_zero_point,
-        reduce_range=config.reduce_range,
     )
 
     setattr(
